+++ css/global.css
@@ -14,7 +14,7 @@ menu, nav, output, ruby, section, summary,
 time, mark, audio, video {
   margin: 0;
   padding: 0;
-  border: 0;
+  border: 1px solid red;
   box-sizing: border-box;
 }
 /* HTML5 display-role reset for older browsers */
@@ -26,10 +26,6 @@ footer, header, hgroup, menu, nav, section {
 html {
   font-size: 100%;
   font-family: Arial, sans-serif;
-}
-
-body {
-   line-height: 1;
 }
 
 ol, ul {
