+++ index.html
@@ -164,7 +164,7 @@
 
           <div class="nav-main middle">
             <div class="nav-departments">
-              <a href="#" class='nav-menu-box'>
+              <a href="#" class='nav-menu-box nav-down-arr'>
                 <div class="nav-menu-text">
                   <span class="text1"></span>
                   <span class="text2">Departments</span>
@@ -181,7 +181,7 @@
           </div>
         </div>
         <div class="nav-plan-layer">
-           <div class="nav-hidden-plan">
+          <div class="nav-hidden-plan hidden-content">
             <div class="plan-container">
               <div class="plan-wrapper">
                 <div class="plan-header">
@@ -261,7 +261,7 @@
               </div>
             </div>
           </div>
-          <div class="nav-hidden-bar">
+          <div class="nav-hidden-bar hidden-content">
             <ul class="bar-wrapper">
               <li class="bar-prime-logo"></li>
               <li class="bar-see-more">
@@ -276,7 +276,7 @@
               </li>
             </ul>
           </div>
-          <div class="nav-hidden-close">
+          <div class="nav-hidden-close hidden-content">
             <span class="close-arrow" id="content-close-btn"></span>
           </div>
         </div>
